requirements.txt
Flask
Flask-Cors
PyJWT
Flask-SQLAlchemy
Werkzeug
